ReactionViewModelをインタフェースと実装に分ける
